# backend/routes/doc/openapi_post_spec.yaml
components:
  schemas:
    Body_create_post_api_posts__post:
      properties:
        comment:
          title: Comment
          type: string
        img_path:
          title: Img Path
          type: string
        title:
          title: Title
          type: string
        user_id:
          title: User Id
          type: integer
      required:
      - user_id
      - title
      - comment
      - img_path
      title: Body_create_post_api_posts__post
      type: object
    Body_update_post_api_posts__post_id__put:
      properties:
        comment:
          anyOf:
          - type: string
          - type: 'null'
          title: Comment
        img_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Img Path
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
      title: Body_update_post_api_posts__post_id__put
      type: object
    Body_user_create_users__post:
      properties:
        email:
          title: Email
          type: string
        name:
          title: Name
          type: string
        surname:
          title: Surname
          type: string
        username:
          title: Username
          type: string
      required:
      - username
      - name
      - surname
      - email
      title: Body_user_create_users__post
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    Post:
      properties:
        comment:
          title: Comment
          type: string
        creation_date:
          format: date-time
          title: Creation Date
          type: string
        id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Id
        img_path:
          title: Img Path
          type: string
        title:
          title: Title
          type: string
        update_date:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Update Date
        user_id:
          title: User Id
          type: integer
      required:
      - title
      - comment
      - img_path
      - user_id
      title: Post
      type: object
    User:
      properties:
        email:
          title: Email
          type: string
        id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Id
        name:
          title: Name
          type: string
        surname:
          title: Surname
          type: string
        username:
          title: Username
          type: string
      required:
      - username
      - name
      - surname
      - email
      title: User
      type: object
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  title: Simple Social Media API
  version: '0.1'
openapi: 3.1.0
paths:
  /posts/:
    get:
      operationId: posts_list_posts__get
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Post'
                title: Response Posts List Posts  Get
                type: array
          description: Successful Response
      summary: Posts List
      tags:
      - Posts
    post:
      operationId: create_post_api_posts__post
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Body_create_post_api_posts__post'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Post Api
      tags:
      - Posts
  /posts/search:
    get:
      operationId: posts_by_query_posts_search_get
      parameters:
      - in: query
        name: query
        required: true
        schema:
          title: Query
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Post'
                title: Response Posts By Query Posts Search Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Posts By Query
      tags:
      - Posts
  /posts/user/{user_id}:
    get:
      operationId: posts_by_user_id_posts_user__user_id__get
      parameters:
      - in: path
        name: user_id
        required: true
        schema:
          title: User Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Post'
                title: Response Posts By User Id Posts User  User Id  Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Posts By User Id
      tags:
      - Posts
  /posts/{post_id}:
    delete:
      operationId: delete_post_by_id_posts__post_id__delete
      parameters:
      - in: path
        name: post_id
        required: true
        schema:
          title: Post Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete Post By Id
      tags:
      - Posts
    get:
      operationId: post_by_id_posts__post_id__get
      parameters:
      - in: path
        name: post_id
        required: true
        schema:
          title: Post Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Post By Id
      tags:
      - Posts
    put:
      operationId: update_post_api_posts__post_id__put
      parameters:
      - in: path
        name: post_id
        required: true
        schema:
          title: Post Id
          type: integer
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Body_update_post_api_posts__post_id__put'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Update Post Api
      tags:
      - Posts
  /users/:
    get:
      operationId: users_list_users__get
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/User'
                title: Response Users List Users  Get
                type: array
          description: Successful Response
      summary: Users List
      tags:
      - Users
    post:
      operationId: user_create_users__post
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Body_user_create_users__post'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: User Create
      tags:
      - Users
  /users/search:
    get:
      operationId: users_by_query_users_search_get
      parameters:
      - in: query
        name: query
        required: true
        schema:
          title: Query
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/User'
                title: Response Users By Query Users Search Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Users By Query
      tags:
      - Users
  /users/username/{username}:
    get:
      operationId: user_by_name_users_username__username__get
      parameters:
      - in: path
        name: username
        required: true
        schema:
          title: Username
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: User By Name
      tags:
      - Users
  /users/{user_id}:
    delete:
      operationId: user_delete_by_id_users__user_id__delete
      parameters:
      - in: path
        name: user_id
        required: true
        schema:
          title: User Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: User Delete By Id
      tags:
      - Users
    get:
      operationId: user_by_id_users__user_id__get
      parameters:
      - in: path
        name: user_id
        required: true
        schema:
          title: User Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: User By Id
      tags:
      - Users